+++ src/NYoutubeDL/Models/VideoInfo.cs
@@ -34,7 +34,7 @@ namespace NYoutubeDL.Models
     {
         public string _type { get; set; }
 
-        public int? abr { get; set; }
+        public double? abr { get; set; }
 
         public string acodec { get; set; }
